Updated rendering to have glowing trails behind the balls. Added level 6.
The balls now have coloured glowing trails which help to give a sense of direction
to them when you're not looking directly at them. This is a bit more CPU load but
worth it, I think.

I also implemented level 6 - seeing as I had a background image for it.

And the intro screenshot has been updated to include the glowing trails.